+++ Tortuga.Chain/Shared/Tests/Core/MetadataTests.cs
@@ -65,35 +65,6 @@ public class MetadataTests : TestBase
 
 #endif
 
-#if SQL_SERVER_MDS
-
-	[DataTestMethod, ViewData(DataSourceGroup.All)]
-	public void ViewIndexes(string dataSourceName, DataSourceType mode, string tableName)
-	{
-		var dataSource = DataSource(dataSourceName, mode);
-		try
-		{
-			var table = dataSource.DatabaseMetadata.GetTableOrView(tableName);
-			var indexes = table.GetIndexes();
-			Assert.IsTrue(indexes.Where(i => i.IsPrimaryKey).Count() <= 1, "No more than one primary key");
-
-			if (table.Columns.Any(c => c.IsPrimaryKey))
-				Assert.IsTrue(indexes.Where(i => i.IsPrimaryKey).Count() <= 1, "A column is marked as primary, so there should be a primary index.");
-
-			foreach (var index in indexes)
-			{
-				//Assert.IsTrue(index.Columns.Count > 0, "Indexes should have columns");
-				Assert.IsFalse(string.IsNullOrWhiteSpace(index.Name), "indexes should have names");
-			}
-		}
-		finally
-		{
-			Release(dataSource);
-		}
-	}
-
-#endif
-
 #if SQL_SERVER_MDS || SQL_SERVER_OLEDB || POSTGRESQL
 
 	[DataTestMethod, BasicData(DataSourceGroup.All)]