+++ Tortuga.Chain/Tortuga.Chain.SqlServer.MDS/SqlServer/SqlServerDataSourceBase.Traits.cs
@@ -28,6 +28,7 @@ namespace Tortuga.Chain.SqlServer;
 [UseTrait(typeof(SupportsProcedureTrait<AbstractCommand, AbstractParameter, AbstractObjectName, AbstractDbType>))]
 [UseTrait(typeof(SupportsTableFunctionTrait<AbstractCommand, AbstractParameter, AbstractObjectName, AbstractDbType, AbstractLimitOption>))]
 [UseTrait(typeof(SupportsGetByColumnListTrait<AbstractCommand, AbstractParameter, AbstractObjectName, AbstractDbType>))]
+[UseTrait(typeof(SupportsDisableIndexesTrait<AbstractParameter, AbstractObjectName, AbstractDbType>))]
 partial class SqlServerDataSourceBase : ICrudDataSource, IAdvancedCrudDataSource
 {
 	DatabaseMetadataCache<AbstractObjectName, AbstractDbType> ICommandHelper<AbstractObjectName, AbstractDbType>.DatabaseMetadata => DatabaseMetadata;
@@ -51,7 +52,6 @@ partial class SqlServerDataSourceBase : ICrudDataSource, IAdvancedCrudDataSource
 		var parameters = new List<SqlParameter>();
 		for (var i = 0; i < keyList.Count; i++)
 		{
-
 			parameters.Add(Utilities.CreateParameter(columnMetadata, "@Param" + i, keyList[i]));
 		}
 
@@ -108,7 +108,6 @@ partial class SqlServerDataSourceBase : ICrudDataSource, IAdvancedCrudDataSource
 
 		return new SqlServerTableOrView<TObject>(this, tableName, where, parameters);
 	}
-
 
 	MultipleRowDbCommandBuilder<AbstractCommand, AbstractParameter> IGetByKeyHelper<AbstractCommand, AbstractParameter, AbstractObjectName, AbstractDbType>.OnGetByKeyList<TObject, TKey>(AbstractObjectName tableName, ColumnMetadata<AbstractDbType> keyColumn, IEnumerable<TKey> keys) where TObject : class
 	{
@@ -221,5 +220,19 @@ where TArgument : class
 		//Verify the table name actually exists.
 		var table = DatabaseMetadata.GetTableOrView(tableName);
 		return Sql("TRUNCATE TABLE " + table.Name.ToQuotedString() + ";").AsNonQuery();
+	}
+
+	private partial ILink<int?> OnEnableIndexes(SqlServerObjectName tableName)
+	{
+		var table = DatabaseMetadata.GetTableOrView(tableName);
+		var sqlStatement = $"ALTER INDEX ALL ON {table.Name.ToQuotedString()} REBUILD";
+		return new SqlServerSqlCall(this, sqlStatement, null).AsNonQuery();
+	}
+
+	private partial ILink<int?> OnDisableIndexes(SqlServerObjectName tableName)
+	{
+		var table = DatabaseMetadata.GetTableOrView(tableName);
+		var sqlStatement = $"ALTER INDEX ALL ON {table.Name.ToQuotedString()} DISABLE";
+		return new SqlServerSqlCall(this, sqlStatement, null).AsNonQuery();
 	}
 }